requirements.txt
opencv-python
numpy
mediapipe
pyyaml
matplotlib
tqdm
streamlit